replace i18n conditions on links with built in command to do the same
