+++ components/Authors/AuthorName.vue
@@ -1,7 +1,7 @@
 <template>
   <fragment>
     <span v-if="author.error">{{ author.name }}</span>
-    <NLink v-else :to="`${$i18n.locale === 'en' ? '' : `/${$i18n.locale}`}/authors/${author.username}`" class="Author__Name">
+    <NLink v-else :to="localePath(`/authors/${author.username}`)" class="Author__Name">
       {{ author.name }}
     </NLink>
   </fragment>
