+++ components/Card.vue
@@ -8,7 +8,7 @@
         <a v-if="post.redirect" class="Vlt-card__content" :href="post.redirect" :title="post.title">
           <h2>{{ post.title | truncate(73, '...') }}</h2>
         </a>
-        <NLink v-else class="Vlt-card__content" :to="post.route" :title="post.title">
+        <NLink v-else class="Vlt-card__content" :to="localePath(post.route)" :title="post.title">
           <h2>{{ post.title | truncate(73, '...') }}</h2>
         </NLink>
         <div class="Vlt-card__footer Vlt-card__footer--short Vlt-card__footer--noborder">
