+++ components/Pagination.vue
@@ -3,19 +3,41 @@
     <ul>
       <li class="Vlt-table__pagination__prev">
         <span v-if="page === first">Previous</span>
-        <nuxt-link v-else :to="`${$i18n.locale === 'en' ? '' : `/${$i18n.locale}`}${route}/p/${page - 1}`">
+        <nuxt-link v-else :to="localePath(`${route}/p/${page - 1}`)">
           Previous
         </nuxt-link>
       </li>
-      <li v-for="p in pages" :key="`page-${p}`" :class="{ 'Vlt-table__pagination__current': p === page }">
-        <span v-if="p === page">{{ p }}</span>
-        <nuxt-link v-else :to="`${$i18n.locale === 'en' ? '' : `/${$i18n.locale}`}${route}/p/${p}`">
-          {{ p }}
+      <li v-if="page-4 > first" class="Vlt-table__pagination__prev">
+        <nuxt-link :to="localePath(`${route}/p/${page - 5}`)">
+          ...
+        </nuxt-link>
+      </li>
+      <template
+        v-for="(p, index) in pages" 
+      >
+        <li
+          v-if="
+            p === page ||
+              (p > (page-5) && p < page) ||
+              (p < (page+5) && p > page)
+          "
+          :key="index"
+          :class="{ 'Vlt-table__pagination__current': p === page }"
+        >
+          <span v-if="p === page">{{ p }}</span>
+          <nuxt-link v-else :to="localePath(`${route}/p/${p}`)">
+            {{ p }}
+          </nuxt-link>
+        </li>
+      </template>
+      <li v-if="page+4 < last" class="Vlt-table__pagination__prev">
+        <nuxt-link :to="localePath(`${route}/p/${page + 5}`)">
+          ...
         </nuxt-link>
       </li>
       <li class="Vlt-table__pagination__next">
         <span v-if="page === last">Next</span>
-        <nuxt-link v-else :to="`${$i18n.locale === 'en' ? '' : `/${$i18n.locale}`}${route}/p/${page + 1}`">
+        <nuxt-link v-else :to="localePath(`${route}/p/${page + 1}`)">
           Next
         </nuxt-link>
       </li>
